Move code for matching platform into the index package
